Implement filter based on a search period
Different search periods to filter posts is one of the things that differentiates each of the current aggregators. Enabling the definition of a search period via configuration data allows reusing code.

A string representing a search period can be defined in the feed generator config. This string can be parsed into a timedelta using a regex and is then used to find the oldest post that will be included in the podcast feed.

A date format string can also be provided in the feed generator configuration. This format string is used to convert the date from the source feed and filter the posts by date.