Create a boolean named `running_on_gcp` which is False when running locally, and True when running on GCP.
I thought about using environment variables for this, but I figured it'd be easier to set up here. The environment variables are more effort to set up or change because you have to do so through the UI (well, maybe there's a way to do it through the command line, but then you have to go learn that), and I figured having the boolean is a clean enough solution anyways so why not. Maybe a little easier for people new to the code to understand.